Clean up external project for packaging
A quirk of the way CMake handles external projects that use git is that
each time the target is evaluated, an update step is performed to make
sure the sources are fresh. Unfortunately, this causes the external
project's build and install steps to be re-evaluated as well, even if
the git update didn't actually change anything.

When this package is processed, it is first built, and then installed
with a separate invocation by the build tool. Because of the quirk, the
external project is re-built during the installation phase. In most
circumstances, the binaries built in the prior invocation are deemed
adequate and no re-compilation is performed.

Unfortunately, this wreaks havoc when the installation invocation uses
DESTDIR, which is the case for the debian and RPM builds on the
buildfarm. The DESTDIR environment variable affects the external
project's installation into the staging directory, and you end up with a
bunch of extra files getting installed by the debian into `/tmp`, which
is certainly not right.

What we've been doing in the other vendor packages in ROS 2 is
suppressing the update step so that CMake doesn't attempt to re-build
the external project during the installation phase at all. This has the
side effect that changing the GIT_TAG doesn't cause the external project
to invoke git to pull a new version of the package, so to work around
that, we've been embedding the version number in the name of the
external project, which will obviously result in a new build when the
target git tag changes.